[joel.jo] 방탈출 인증 관리 step1 & step2 (#18)
* feat: 토큰 발급 API 구현

* feat: 회원정보 조회 API 구현

* docs: 구현할 기능목록 최신화

* feat: 비로그인 사용자 제어 로직 구현

* feat: 예약 삭제 기능 구현

* feat: ArgumentResolver 등록

* style: 코드 포맷팅

* test: 예약 E2E 테스트 작성

* feat: 토큰 부재시 401 에러 띄우도록 구현

* test: 토큰이 필요한 테스트에 토큰 추가

* refactor: ExceptionHandler 한 곳에서 관리하도록 수정

* style: 코드 포맷팅

* refactor: 예외가 메시지를 가지도록 수정

* test: 본인 정보 가져오는 테스트 작성

* style: 코드 포맷팅

* refactor: MemberService가 JwtToken 생성의 책임을 가지도록 수정

* refactor: Member name에 unique 제약조건 추가

* refactor: 사용자 이름이 고유한지 검증하는 로직 추가

* refactor: 2개 이상의 쿼리를 날리는 Service 메서드에 Transactional 어노테이션 추가

+++ src/main/java/nextstep/reservation/ReservationService.java
@@ -1,27 +1,36 @@
 package nextstep.reservation;
 
+import nextstep.member.Member;
+import nextstep.member.MemberDao;
 import nextstep.schedule.Schedule;
 import nextstep.schedule.ScheduleDao;
+import nextstep.support.AuthorizationException;
 import nextstep.support.DuplicateEntityException;
+import nextstep.support.NotExistEntityException;
 import nextstep.theme.Theme;
 import nextstep.theme.ThemeDao;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Objects;
 
 @Service
 public class ReservationService {
     public final ReservationDao reservationDao;
     public final ThemeDao themeDao;
     public final ScheduleDao scheduleDao;
+    public final MemberDao memberDao;
 
-    public ReservationService(ReservationDao reservationDao, ThemeDao themeDao, ScheduleDao scheduleDao) {
+    public ReservationService(ReservationDao reservationDao, ThemeDao themeDao, ScheduleDao scheduleDao, MemberDao memberDao) {
         this.reservationDao = reservationDao;
         this.themeDao = themeDao;
         this.scheduleDao = scheduleDao;
+        this.memberDao = memberDao;
     }
 
-    public Long create(ReservationRequest reservationRequest) {
+    @Transactional
+    public Long create(String username, ReservationRequest reservationRequest) {
         Schedule schedule = scheduleDao.findById(reservationRequest.getScheduleId());
         if (schedule == null) {
             throw new NullPointerException();
@@ -32,14 +41,17 @@ public class ReservationService {
             throw new DuplicateEntityException();
         }
 
+        Member member = memberDao.findByUsername(username);
+
         Reservation newReservation = new Reservation(
                 schedule,
-                reservationRequest.getName()
+                member.getName()
         );
 
         return reservationDao.save(newReservation);
     }
 
+    @Transactional
     public List<Reservation> findAllByThemeIdAndDate(Long themeId, String date) {
         Theme theme = themeDao.findById(themeId);
         if (theme == null) {
@@ -49,10 +61,16 @@ public class ReservationService {
         return reservationDao.findAllByThemeIdAndDate(themeId, date);
     }
 
-    public void deleteById(Long id) {
+    @Transactional
+    public void deleteById(String username, Long id) {
         Reservation reservation = reservationDao.findById(id);
-        if (reservation == null) {
-            throw new NullPointerException();
+        Member member = memberDao.findByUsername(username);
+        if (reservation == null || member == null) {
+            throw new NotExistEntityException();
+        }
+
+        if (!Objects.equals(member.getName(), reservation.getName())) {
+            throw new AuthorizationException();
         }
 
         reservationDao.deleteById(id);
